Add EditorState Race Conditions documentation
Summary:
**Summary**
This new documentation page is under "Advanced Topics" and between "Text Direction" and "Issues and Pitfalls".
It documents the ways to have race conditions based on consecutive EditorState updates and best practices to avoid this issue.
This was motivated by issue #318.

It also updates the CSS to make all images' width to 650px which is the width of the right column content, hope that's fine. (this version of markdown (Marked) doesn't allow to set the image width).

**Test Plan**
Go to /docs/advanced-topics-editorstate-race-conditions.html and check the page loads.

Screenshot:
![localhost-8080-docs-advanced-topics-editorstate-race-conditions html](https://user-images.githubusercontent.com/78122/30224834-1b445976-9485-11e7-9387-d74c55f00621.png)
Closes https://github.com/facebook/draft-js/pull/1371

Differential Revision: D5814889

fbshipit-source-id: 4faee30e456f335d07e2fcc98169c242e6228a29