Exploration - Feature tree / Rendering tree data format
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Rendering tree data format**

This PR adds support for react components to render the tree data structure

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1508

Differential Revision: D6341091

fbshipit-source-id: e285bb04ca51eb55b75c6912a2f5d283a4145d2d