Exploration - Feature tree / Adding regression tests
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Adding regression tests**

This PR adds regression tests to make sure we do not enable experimental features unless the dark feature flag is enabled

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1527

Differential Revision: D6399454

fbshipit-source-id: 60c020eaf57c73de9ee5bf7596581fb25678f009