Fix text node key logic in async mode
Summary: Previously if componentWillUpdate runs twice (negating its effect), then typing "hi", Cmd-A, "hello" would end you up with a mash of text instead of "hello". This diff changes it to only flip the flag once per commit which should be the relevant time. This way each time it renders (and sCU returns true), we do successfully recreate the DOM.

Reviewed By: flarnie

Differential Revision: D6431545

fbshipit-source-id: 9400e24bc1ffddf2946cbfea5cb6fb1d8fbe4a5c