Exploration - Feature tree / Adding support for text into tree nodes
Summary:
This PR is part of a series of PR's that will be exploring **tree data block support** in Draft.

**Adding support for text into tree nodes**

This PR adds support for converting pasted text into tree block nodes

***

**Note:**
This is unstable and not part of the public API and should not be used by
production systems.
Closes https://github.com/facebook/draft-js/pull/1541

Differential Revision: D6457757

fbshipit-source-id: 527e54929a3e8f4c68141fd3bd260066fed14977