Workaround for BlockNode variance issue caused by flow transform (#1621)
Summary:
**Summary**

This should fix #1621 (it does in our project).

This manually transforms the BlockNode interface to use properties like the build step will do, but it adds covariant property annotations to match the expectation that classes have read-only methods. A better solution would be to fix the transform itself to do the right thing with interface methods, but this should be equivalent and is easier (for me) to do.

**Test Plan**

Confirm that the output lib/BlockNode.js.flow includes the covariant property annotations (`+`) before each of the properties that were previously methods. The automated test infrastructure only runs over src so it's harder to verify in travis as-is.
Closes https://github.com/facebook/draft-js/pull/1683

Differential Revision: D7399839

fbshipit-source-id: 16fb390789f2ab7cff3781d854c6d9af900de906