1/n splitting PR #1828: Initial forking of RichTextEditorUtil (#1828)
Summary:
We are splitting up and updating pieces of this PR by Mitermayer.

This chunk forks the RichTextEditorUtil and puts them both under the same Flow
interface, also adding a test. The forked NestedRichTextEditorUtil is a WIP
version of this utility.

This is directly taken from mitermayer 's fork, props to him
Pull Request resolved: https://github.com/facebook/draft-js/pull/1828

Reviewed By: mitermayer

Differential Revision: D9181952

Pulled By: mitermayer

fbshipit-source-id: 75f08fdb228ae1b45f97ae29aeae2f3d5ea53650