Fixes warning for missing keys in example color controls (#1853)
Summary:
**Summary**

Fixes a warning in the color examples for missing keys for elements in an array for the style controls.

```
react.development.js:336 Warning: Each child in an array or iterator should have a unique "key" prop.

Check the render method of `ColorControls`. See https://fb.me/react-warning-keys for more information.
    in StyleButton (created by ColorControls)
    in ColorControls (created by ColorfulEditorExample)
    in div (created by ColorfulEditorExample)
    in ColorfulEditorExample
```

**Test Plan**

Manual tests on Chrome, warnings are gone

**Note:** This PR also fixes the [Travis CI build failure](https://travis-ci.org/facebook/draft-js/jobs/424170309) introduced with e6c693c
Pull Request resolved: https://github.com/facebook/draft-js/pull/1853

Differential Revision: D9654616

fbshipit-source-id: 1e4e9a4ef2eb896306404acc31623076e107b91c