Fix leading line feed conversion
Summary:
Pasting text from Quip results in a space being inserted at the beginning of each new line (except for the first one), due to blocks including \n as a leading character that is then turned into a space.

I couldn't repro the issue with Word flagged in the comment (could be Word on Windows, or one of the older versions of Word?), but just in case I'm only removing the leading \n that is invisible in HTML and leaving the other replacements as is.

Reviewed By: claudiopro

Differential Revision: D14199824

fbshipit-source-id: c6b9cfa72e3f75d9fae485bb87fb5bd40862375a