catch errors when encoding entity map
Summary:
We've had several reports of people's content failing to save in our product. The only error message we have is this (with no callstack, unfortunately):
```
Unknown DraftEntity key: undefined.
```
I poked around in the DraftJS code, and my guess is that this is due to some character having an entity key associated with it that for whatever reason does not appear in the entity map. This diff "fixes" the issue by not throwing an error in this situation.

This is definitely not as good as figuring out the root cause of the entity mismatch, but it seems reasonable for `convertFromDraftStateToRaw` to work with slightly malformed input.

Reviewed By: niveditc

Differential Revision: D16421104

fbshipit-source-id: a2046ad0cb166dac97187fd3c524f9c8cb4017ea