Short-circuit getUpdatedSelectionState for invalid selection updates on prod
Summary:
The code already did this for __DEV__, but it was misswritten, since instead of skipping the wanring on prod, it skipped the whole check.

This essentially deals with empty-string keys, which mean nothing.

This wasn't much of an issue— we'd simply not get any leafs and exit on a following check.

Reviewed By: claudiopro

Differential Revision: D21145150

fbshipit-source-id: bb78aa076d98f5d642a2ae8eb13c158dde8a76e3