Remove loadWithEntities, add setEntityMap
Summary:
Deprecating DraftEntity will make entityMap part of the immutable ContentState data structure, so side effects wont happen.

This means replacing the entityMap will return a new content state, not change the current one.

Let's change this function so it mimicks this behaviour now, so we don't have to do so later.

I also rename it to reflect this change in behaviour. This revealed the call-sites of the old one.

I added some TODOs to places where I noticed this function was used to clear the global state when the `ContentState` was supposed to be empty/cleared too, which is a workaround that wont be needed once we remove `DraftEntity` so there's no more global state— everything will live in `ContentState`.

shalabhvyas please check that I do thinks properly in `CMSWYSIWYGEditorCoEditingMutator`

Reviewed By: shalabhvyas

Differential Revision: D24069524

fbshipit-source-id: c8f21489bb17537e6366db0b6d57a903a2f5fe3c