convertFromRawToDraftState creates proper entity maps.
Summary:
D24174087 had a broken e2e test which I traced to `convertFromRawToDraftState`. Turns out the file passes around an `entityMap` all over the place, but it's not acutally an `EntityMap`. It is a map between IDs— one is the `key` used in the enconding and the other is the actual entity ID once its added to the content state.

Unfortunatley, the file was littered with `*` types which since like 2018 (when they were deprecated) are essentially aliases of `any`. Since this "key map" was passed around being called `entityMap`, in the end our content state was built with it.

Of course a random object is not a valid entity map. This became aparent when D24174087 went a bit further and expected EntityMap to be an `Immutable.OrderedMap`. Suddenly, we noticed it was just a random object and things failed.

While I was at it I improved some flow types and flowified the tests.

Reviewed By: AjkP

Differential Revision: D24325705

fbshipit-source-id: bb8bf1b04b3564144bd6186cfb6598d6a511a4b9