Flowify getWindowForNode
Summary:
Thought it'd take a minute or two to look at some Quality Analizer tasks and see if maybe there was some simple things to flowify in Draft.js. I found me this function.

NOTE: the libdef for Window is internal, so I aded one for the external repo. There really isn't a simple way of using the external one for the Draft codebase, internally, so I made the external type of Window just be any (which is the default anyway). This effectivley increases type saftey internally only, but it's something ¯\_(ツ)_/¯

Reviewed By: AjkP

Differential Revision: D24934095

fbshipit-source-id: bfa12f5e26a79afba7c71b868f26f4da8559a62e