Fix the focus ring of the composer outline
Summary:
The issue can be reproduced with tools which force the outline ring to show up.

When there are 5 or more lines of text in the chat composer, a scrollbar appears.
The focus ring of the editor moves together with this scrollbar, which creates a very weird visual effect.
This changes fixes that.

Reviewed By: andrewdacenko

Differential Revision: D26689495

fbshipit-source-id: 3fbac77e223ce8ace0025852eb6f0d023bac328c